simplify testing setup
