--- contracts/TransformImport.sol
@@ -1,1 +1,0 @@
-../test/solc-0.6/contracts/Local.sol
