--- contracts/solc-0.6
@@ -1,1 +1,0 @@
-../test/solc-0.6/contracts
