+++ contracts/solc-0.6/AbstractContract.sol
