+++ contracts/solc-0.6/ClassInheritance.sol
