+++ contracts/solc-0.6/DiamondInheritance.sol
