+++ contracts/solc-0.6/ElementaryTypesWithConstructor.sol
