+++ contracts/solc-0.6/Interface.sol
