+++ contracts/solc-0.6/SimpleInheritance.sol
