+++ contracts/solc-0.6/StringConstructor.sol
