+++ src/index.test.ts
@@ -1,0 +1,53 @@
+import _test, { TestInterface } from 'ava';
+import hre from 'hardhat';
+
+import { getBuildInfo } from './test-utils/get-build-info';
+
+import { OutputFile, transpile } from '.';
+import { SolcInput, SolcOutput } from './solc/input-output';
+
+const test = _test as TestInterface<Context>;
+
+interface Context {
+  files: OutputFile[];
+}
+
+const fileNames = [
+  'ClassInheritance.sol',
+  'Override.sol',
+  'DiamondInheritance.sol',
+  'Deep.sol',
+  'ElementaryTypes.sol',
+  'ElementaryTypesWithConstructor.sol',
+  'Imported.sol',
+  'Local.sol',
+  'SimpleInheritance.sol',
+  'StringConstructor.sol',
+  'Library.sol',
+  'AbstractContract.sol',
+  'Interface.sol',
+  'Rename.sol',
+];
+
+test.serial.before('compile', async t => {
+  const buildInfo = await getBuildInfo('0.6');
+  const solcInput = buildInfo.input;
+  const solcOutput = buildInfo.output as SolcOutput;
+
+  console.log(Object.keys(solcInput.sources));
+
+  t.context.files = await transpile(solcInput, solcOutput, hre.config.paths);
+});
+
+for (const fileName of fileNames) {
+  test(fileName, t => {
+    const file = t.context.files.find(f => f.fileName === fileName);
+    t.not(file, undefined, 'file not found');
+    t.snapshot(file);
+  });
+}
+
+test('AlreadyUpgradeable.sol', t => {
+  const file = t.context.files.find(f => f.fileName === 'AlreadyUpgradeable.sol');
+  t.is(file, undefined);
+});
