+++ src/index.test.ts.md
@@ -1,6 +1,6 @@
-# Snapshot report for `test/solc-0.6/test.js`
+# Snapshot report for `src/index.test.ts`
 
-The actual snapshot is saved in `test.js.snap`.
+The actual snapshot is saved in `index.test.ts.snap`.
 
 Generated by [AVA](https://avajs.dev).
 
@@ -10,9 +10,9 @@ Generated by [AVA](https://avajs.dev).
 
     {
       fileName: 'ClassInheritance.sol',
-      path: 'contracts/ClassInheritanceUpgradeable.sol',
+      path: 'contracts/solc-0.6/ClassInheritanceUpgradeable.sol',
       source: `pragma solidity ^0.6.0;␊
-      import "./Initializable.sol";␊
+      import "../Initializable.sol";␊
       ␊
       contract CIAUpgradeable is Initializable {␊
           uint256 public foo;␊
@@ -49,9 +49,9 @@ Generated by [AVA](https://avajs.dev).
 
     {
       fileName: 'Override.sol',
-      path: 'contracts/OverrideUpgradeable.sol',
+      path: 'contracts/solc-0.6/OverrideUpgradeable.sol',
       source: `pragma solidity ^0.6.0;␊
-      import "./Initializable.sol";␊
+      import "../Initializable.sol";␊
       ␊
       contract Parent1Upgradeable is Initializable {␊
           function __Parent1_init() internal onlyInitializing {␊
@@ -105,9 +105,9 @@ Generated by [AVA](https://avajs.dev).
 
     {
       fileName: 'DiamondInheritance.sol',
-      path: 'contracts/DiamondInheritanceUpgradeable.sol',
+      path: 'contracts/solc-0.6/DiamondInheritanceUpgradeable.sol',
       source: `pragma solidity ^0.6.0;␊
-      import "./Initializable.sol";␊
+      import "../Initializable.sol";␊
       ␊
       contract DAUpgradeable is Initializable {␊
           event Log(string);␊
@@ -175,11 +175,11 @@ Generated by [AVA](https://avajs.dev).
 
     {
       fileName: 'Deep.sol',
-      path: 'contracts/one/two/three/DeepUpgradeable.sol',
+      path: 'contracts/solc-0.6/one/two/three/DeepUpgradeable.sol',
       source: `pragma solidity ^0.6.0;␊
       ␊
       import "../../../ElementaryTypesUpgradeable.sol";␊
-      import "../../../Initializable.sol";␊
+      import "../../../../Initializable.sol";␊
       ␊
       contract DeepUpgradeable is Initializable {␊
           function __Deep_init() internal onlyInitializing {␊
@@ -199,9 +199,9 @@ Generated by [AVA](https://avajs.dev).
 
     {
       fileName: 'ElementaryTypes.sol',
-      path: 'contracts/ElementaryTypesUpgradeable.sol',
+      path: 'contracts/solc-0.6/ElementaryTypesUpgradeable.sol',
       source: `pragma solidity ^0.6.0;␊
-      import "./Initializable.sol";␊
+      import "../Initializable.sol";␊
       ␊
       contract ElementaryTypesUpgradeable is Initializable  {␊
           function __ElementaryTypes_init() internal onlyInitializing {␊
@@ -237,9 +237,9 @@ Generated by [AVA](https://avajs.dev).
 
     {
       fileName: 'ElementaryTypesWithConstructor.sol',
-      path: 'contracts/ElementaryTypesWithConstructorUpgradeable.sol',
+      path: 'contracts/solc-0.6/ElementaryTypesWithConstructorUpgradeable.sol',
       source: `pragma solidity ^0.6.0;␊
-      import "./Initializable.sol";␊
+      import "../Initializable.sol";␊
       ␊
       contract ElementaryTypesWithConstructorUpgradeable is Initializable  {␊
           address public owner;␊
@@ -272,9 +272,9 @@ Generated by [AVA](https://avajs.dev).
 
     {
       fileName: 'Imported.sol',
-      path: 'contracts/ImportedUpgradeable.sol',
+      path: 'contracts/solc-0.6/ImportedUpgradeable.sol',
       source: `pragma solidity ^0.6.0;␊
-      import "./Initializable.sol";␊
+      import "../Initializable.sol";␊
       ␊
       contract Imported1Upgradeable is Initializable {␊
           function __Imported1_init(uint256 x, uint256 y) internal onlyInitializing {␊
@@ -303,11 +303,11 @@ Generated by [AVA](https://avajs.dev).
 
     {
       fileName: 'Local.sol',
-      path: 'contracts/LocalUpgradeable.sol',
+      path: 'contracts/solc-0.6/LocalUpgradeable.sol',
       source: `pragma solidity ^0.6.0;␊
       ␊
       import "./ImportedUpgradeable.sol";␊
-      import "./Initializable.sol";␊
+      import "../Initializable.sol";␊
       ␊
       contract LocalUpgradeable is Initializable, Imported2Upgradeable {␊
           function __Local_init(uint x, uint y) internal onlyInitializing {␊
@@ -328,9 +328,9 @@ Generated by [AVA](https://avajs.dev).
 
     {
       fileName: 'SimpleInheritance.sol',
-      path: 'contracts/SimpleInheritanceUpgradeable.sol',
+      path: 'contracts/solc-0.6/SimpleInheritanceUpgradeable.sol',
       source: `pragma solidity ^0.6.0;␊
-      import "./Initializable.sol";␊
+      import "../Initializable.sol";␊
       ␊
       contract SIAUpgradeable is Initializable {␊
           uint256 public foo;␊
@@ -382,9 +382,9 @@ Generated by [AVA](https://avajs.dev).
 
     {
       fileName: 'StringConstructor.sol',
-      path: 'contracts/StringConstructorUpgradeable.sol',
+      path: 'contracts/solc-0.6/StringConstructorUpgradeable.sol',
       source: `pragma solidity ^0.6.0;␊
-      import "./Initializable.sol";␊
+      import "../Initializable.sol";␊
       ␊
       contract StringConstructorUpgradeable is Initializable {␊
           function __StringConstructor_init(string memory message) internal onlyInitializing {␊
@@ -405,7 +405,7 @@ Generated by [AVA](https://avajs.dev).
 
     {
       fileName: 'Library.sol',
-      path: 'contracts/LibraryUpgradeable.sol',
+      path: 'contracts/solc-0.6/LibraryUpgradeable.sol',
       source: `pragma solidity ^0.6.0;␊
       ␊
       library LibraryUpgradeable {␊
@@ -419,9 +419,9 @@ Generated by [AVA](https://avajs.dev).
 
     {
       fileName: 'AbstractContract.sol',
-      path: 'contracts/AbstractContractUpgradeable.sol',
+      path: 'contracts/solc-0.6/AbstractContractUpgradeable.sol',
       source: `pragma solidity ^0.6.0;␊
-      import "./Initializable.sol";␊
+      import "../Initializable.sol";␊
       ␊
       abstract contract AbstractContractUpgradeable is Initializable {␊
           function __AbstractContract_init() internal onlyInitializing {␊
@@ -441,7 +441,7 @@ Generated by [AVA](https://avajs.dev).
 
     {
       fileName: 'Interface.sol',
-      path: 'contracts/InterfaceUpgradeable.sol',
+      path: 'contracts/solc-0.6/InterfaceUpgradeable.sol',
       source: `pragma solidity ^0.6.0;␊
       ␊
       interface InterfaceUpgradeable {␊
@@ -455,9 +455,9 @@ Generated by [AVA](https://avajs.dev).
 
     {
       fileName: 'Rename.sol',
-      path: 'contracts/RenameUpgradeable.sol',
+      path: 'contracts/solc-0.6/RenameUpgradeable.sol',
       source: `pragma solidity ^0.6.0;␊
-      import "./Initializable.sol";␊
+      import "../Initializable.sol";␊
       ␊
       library RenameLibraryUpgradeable {␊
           function test() external {␊
