--- test/solc-0.6/.gitignore
@@ -1,2 +1,0 @@
-/build
-/contracts/__upgradeable__
